+++ frontend/src/components/channels/RefreshChannelButton.tsx
@@ -1,12 +1,12 @@
 import { ArrowClockwise, CheckCircleFill, XOctagonFill } from "react-bootstrap-icons";
-import { PullChannelRequest } from "#/components/events/PullChannelRequest";
+import { PullChannelsRequest } from "#/components/events/PullChannelsRequest";
 import { useAppDispatch, useAppSelector } from "#/store/hooks";
-import { pullChannelRequested } from "#/store/slices/pullChannel";
+import { pullChannelRequested } from "#/store/slices/pullChannels";
 import type { ChannelProps } from "#/types";
 
 const RefreshChannelButton = ({ channel }: ChannelProps) => {
   const dispatch = useAppDispatch();
-  const pullChannel = useAppSelector((state) => state.pullChannel);
+  const pullChannel = useAppSelector((state) => state.pullChannels);
 
   const isPulledRecently = pullChannel.responses[channel.handle] !== undefined;
   const error = pullChannel.errors[channel.handle];
@@ -32,7 +32,7 @@ const RefreshChannelButton = ({ channel }: ChannelProps) => {
   return (
     <>
       {icon}
-      <PullChannelRequest />
+      <PullChannelsRequest />
     </>
   );
 };