+++ frontend/src/components/events/PullChannelsRequest.tsx
@@ -2,32 +2,34 @@ import { usePullMutation } from "#/api";
 import { PullEventsSource } from "#/components/events/PullEventsSource";
 import { useAppDispatch, useAppSelector } from "#/store/hooks";
 import { invalidateChannels } from "#/store/slices/backend";
-import { pullChannelActive, pullChannelError, pullChannelResponse } from "#/store/slices/pullChannel";
+import { invalidateFetchLimits } from "#/store/slices/limits";
+import { pullChannelActive, pullChannelError, pullChannelResponse } from "#/store/slices/pullChannels";
 import type { PullChannelsEvent, PullOperationEvent } from "#/components/events/types";
+import type { PullOperation } from "#/types";
 
-const PullChannelRequest = () => {
+const PullChannelsRequest = () => {
   const dispatch = useAppDispatch();
-  const { active, requested = "" } = useAppSelector((state) => state.pullChannel);
+  const { active, requested } = useAppSelector((state) => state.pullChannels);
 
   const [pull] = usePullMutation();
 
   const requestPullChannel = (eventSubscriptionId: string) => {
-    void pull({
-      body: {
-        event_subscription_id: eventSubscriptionId,
-        operations: [{ pull: "channel", channel_handle: requested }],
-      },
-    });
+    const operations: PullOperation[] = requested.map((handle) => ({ pull: "channel", channel_handle: handle }));
+    void pull({ body: { event_subscription_id: eventSubscriptionId, operations } });
   };
 
   const handlePullOperationEvent = (event: PullOperationEvent) => {
-    dispatch(pullChannelResponse({ channelHandle: requested, status: event.operation.status }));
-    if (event.error) {
-      dispatch(pullChannelError({ channelHandle: requested, error: event.error }));
+    if (event.operation.pull === "channel") {
+      const channelHandle = event.operation.channel_handle;
+      dispatch(pullChannelResponse({ channelHandle, status: event.operation.status }));
+      if (event.error) {
+        dispatch(pullChannelError({ channelHandle, error: event.error }));
+      }
     }
   };
 
   const handlePullChannelsEvent = (event: PullChannelsEvent) => {
+    dispatch(invalidateFetchLimits());
     if (event.channels.length > 0) {
       dispatch(invalidateChannels());
     }
@@ -49,4 +51,4 @@ const PullChannelRequest = () => {
   );
 };
 
-export { PullChannelRequest };
+export { PullChannelsRequest };