+++ frontend/src/store/slices/pullChannels.ts
@@ -1,4 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
+import { isArray } from "lodash";
 import type { Problem } from "#/api";
 import type { PullOperationStatus } from "#/types";
 import type { PayloadAction } from "@reduxjs/toolkit";
@@ -6,27 +7,28 @@ import type { Dictionary } from "lodash";
 
 type PullChannelState = {
   active: boolean;
-  requested?: string;
+  requested: string[];
   responses: Partial<Dictionary<PullOperationStatus>>;
   errors: Partial<Dictionary<Problem>>;
 };
 
 const initialState: PullChannelState = {
   active: false,
+  requested: [],
   responses: {},
   errors: {},
 };
 
-const pullChannel = createSlice({
-  name: "pullChannel",
+const pullChannels = createSlice({
+  name: "pullChannels",
   initialState,
   reducers: {
     pullChannelActive: (state, { payload }: PayloadAction<boolean>) => {
       state.active = payload;
     },
-    pullChannelRequested: (state, { payload }: PayloadAction<string>) => {
+    pullChannelRequested: (state, { payload }: PayloadAction<string | string[]>) => {
       state.active = true;
-      state.requested = payload;
+      state.requested = isArray(payload) ? payload : [payload];
     },
     pullChannelResponse: (
       state,
@@ -37,10 +39,19 @@ const pullChannel = createSlice({
     pullChannelError: (state, { payload }: PayloadAction<{ channelHandle: string; error: Problem }>) => {
       state.errors[payload.channelHandle] = payload.error;
     },
+    pullChannelReset: () => initialState,
   },
 });
 
-const pullChannelReducer = pullChannel.reducer;
-const { pullChannelActive, pullChannelError, pullChannelRequested, pullChannelResponse } = pullChannel.actions;
+const pullChannelsReducer = pullChannels.reducer;
+const { pullChannelActive, pullChannelError, pullChannelRequested, pullChannelReset, pullChannelResponse } =
+  pullChannels.actions;
 
-export { pullChannelActive, pullChannelError, pullChannelReducer, pullChannelRequested, pullChannelResponse };
+export {
+  pullChannelActive,
+  pullChannelError,
+  pullChannelRequested,
+  pullChannelReset,
+  pullChannelResponse,
+  pullChannelsReducer,
+};